services/ocr/pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datalab-python-sdk>=0.3.0",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
]
